chore: allow accessible-error on all roles
